+++ src/components/FormProvider/Context.js
@@ -1,0 +1,6 @@
+import { createContext } from 'react'
+
+const FormContext = createContext()
+FormContext.displayName = 'FormContext'
+
+export default FormContext