+++ src/components/FormProvider/useFormProvider/index.js
@@ -1,0 +1,1 @@
+export { default as useFormProvider } from './useFormProvider'