+++ src/components/ViewProvider/Context.js
@@ -1,6 +1,0 @@
-import { createContext } from 'react'
-
-const ViewContext = createContext()
-ViewContext.displayName = 'ViewContext'
-
-export default ViewContext